Add doc input `Password::class`.
